[fix] error code

+++ Bronze3/Day17/BOJ15894.kt
@@ -5,7 +5,7 @@ import java.io.InputStreamReader
 
 fun main() {
     val br = BufferedReader(InputStreamReader(System.`in`))
-    val n = br.readLine().toInt()
+    val n = br.readLine().toLong()
     println("${n * 4}")
     br.close()
 }
